Refactor by creating a GCS Uploader class to be used in photo and video service
